Extension method restructure
