+++ Olive.Entities.Data/Database/Database.Delete.cs
