+++ Olive.Entities.Data/Extensions/ConcurrencyDataException.cs
