+++ Olive.Entities.Data/QueryExecution/DatabaseQuery.TEntity.Aggregate.cs
