+++ Olive.Mvc/Extentions/Extensions.HtmlHelpers.cs
@@ -22,31 +22,6 @@ namespace Olive.Mvc
     {
         const int DEFAULT_VISIBLE_PAGES = 7;
 
-        // public static UrlHelper GetUrlHelper(this IHtmlHelper html) => new UrlHelper(html.ViewContext.RequestContext);
-
-        // /// <summary>
-        // /// Renders the specified partial view as an HTML-encoded string.
-        // /// </summary>
-        // /// <param name="html">The HTML helper instance that this method extends.</param>
-        // /// <param name="partialViewName">The name of the partial view to render.</param>
-        // /// <param name="model">The view model for the partial view.</param>
-        // /// <returns>The partial view that is rendered as an HTML-encoded string.</returns>
-        // public static HtmlString Partial<T>(this IHtmlHelper html, string partialViewName, T model, bool skipAjaxPost) where T : IViewModel
-        // {
-        //    var request = HttpContext.Current.Request;
-        //    if (skipAjaxPost && request.IsAjaxCall() && request.IsPost()) return HtmlString.Empty;
-
-        //    if (model == null)
-        //    {
-        //        model = (html.ViewContext.Controller as Controller).Bind<T>();
-
-        //        if (model == null)
-        //            throw new Exception("The model object passed to Partial() cannot be null.");
-        //    }
-
-        //    return html.Partial(partialViewName, model);
-        // }
-
         public static HtmlString ToJson(this IHtmlHelper html, object obj)
         {
             if (obj == null) return new HtmlString("[]");
@@ -208,29 +183,6 @@ namespace Olive.Mvc
         public static HtmlString Pagination(this IHtmlHelper html, ListPagination paging, int visiblePages, object htmlAttributes = null, string prefix = null) =>
             new PaginationRenderer(html, paging, visiblePages, htmlAttributes, prefix).Render();
 
-        // public static void RenderAction<TController>(this IHtmlHelper html, string action = "Index")
-        // {
-        //    html.RenderAction(action, typeof(TController).Name.TrimEnd("Controller"));
-        // }
-
-        // /// <summary>
-        // /// Invokes the Index action method of the specified controller and returns the result as an HTML string.
-        // /// <param name="queryParameters">An anonymous object containing query string / route values to pass.</param>
-        // /// </summary>
-        // public static HtmlString Action<TController>(this IHtmlHelper html, object queryParameters)
-        // {
-        //    return Action<TController>(html, "Index", queryParameters);
-        // }
-
-        // /// <summary>
-        // /// Invokes the specified child action method of the specified controller and returns the result as an HTML string.
-        // /// <param name="queryParameters">An anonymous object containing query string / route values to pass.</param>
-        // /// </summary>
-        // public static HtmlString Action<TController>(this IHtmlHelper html, string action = "Index", object queryParameters = null)
-        // {
-        //    return html.Action(action, typeof(TController).Name.TrimEnd("Controller"), queryParameters);
-        // }
-
         /// <summary>
         /// Will join this with other Mvc Html String items;
         /// </summary>
@@ -290,19 +242,6 @@ namespace Olive.Mvc
 
             return new HtmlString(WebTestManager.GetWebTestWidgetHtml(Context.Http.Request));
         }
-
-        // /// <summary>
-        // /// Creates a new Html helper for a new ViewModel object.
-        // /// This enables to start from a new view context, so that normal Html helper methods (such as TextBoxFor, etc) yield the correct name attributes, use correct existing value, etc.
-        // /// </summary>
-        // public static HtmlHelper<TTarget> For<TSource, TTarget>(this IHtmlHelper<TSource> html, TTarget model)
-        // {
-        //    var container = new BasicViewDataContainer { ViewData = new ViewDataDictionary { Model = model } };
-
-        //    var viewContext = new ViewContext(html.ViewContext, html.ViewContext.View, container.ViewData, html.ViewContext.Writer);
-
-        //    return new HtmlHelper<TTarget>(viewContext, container);
-        // }
 
         public static HtmlString RunJavascript(this IHtmlHelper html, string script, PageLifecycleStage stage = PageLifecycleStage.Init) =>
             RunJavascript(html, script, script, stage);
