--- Olive.Web/UrlRewriting/IWebResource.cs
@@ -1,9 +1,0 @@
-﻿using Olive.Entities;
-
-namespace Olive.Web
-{
-    public interface IWebResource : IEntity
-    {
-        string GetUrl();
-    }
-}
