--- Services/Olive.Services.Integration/IIntegrationQueueItem.cs
@@ -1,47 +1,0 @@
-﻿using System;
-using Olive.Entities;
-
-namespace Olive.Services.Integration
-{
-    /// <summary>
-    /// Stores one item requested for processing via an external service.
-    /// </summary>
-    public interface IIntegrationQueueItem : IEntity<Guid>
-    {
-        /// <summary>
-        /// The name of the integration service used for processing this item.
-        /// </summary>        
-        string IntegrationService { get; set; }
-
-        /// <summary>
-        /// The body of the request being sent to the integration service.
-        /// </summary>
-        string Request { get; set; }
-
-        /// <summary>
-        /// The response generated by the remote service.
-        /// </summary>
-        string Response { get; set; }
-
-        /// <summary>
-        /// The error encountered in processing this item.
-        /// </summary>
-        string ErrorResponse { get; set; }
-
-        /// <summary>
-        /// Specifies the date and time when this item was picked for processing.
-        /// If it's null, it means this item has not been picked yet.
-        /// </summary>
-        DateTime? DatePicked { get; set; }
-
-        /// <summary>
-        /// Specifies the date and time when this item was first requested and added to the queue.
-        /// </summary>
-        DateTime RequestDate { get; set; }
-
-        /// <summary>
-        /// Specifies the date when the response was retrieved from the remote service.
-        /// </summary>
-        DateTime? ResponseDate { get; set; }
-    }
-}
