--- Services/Olive.Services.Integration/Olive.Services.Integration.csproj
@@ -1,27 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netcoreapp2.0</TargetFramework>
-    <RootNamespace>Olive.Services.Integration</RootNamespace>
-  </PropertyGroup>
-
-  <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|AnyCPU'">
-    <OutputPath>..\..\@Assemblies\</OutputPath>
-    <DocumentationFile>..\..\@Assemblies\netcoreapp2.0\Olive.Services.Integration.xml</DocumentationFile>
-    <NoWarn>1701;1702;1705;1591;1573</NoWarn>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Newtonsoft.Json" Version="10.0.1" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <Reference Include="Olive">
-      <HintPath>..\..\@Assemblies\netcoreapp2.0\Olive.dll</HintPath>
-    </Reference>
-    <Reference Include="Olive.Entities">
-      <HintPath>..\..\@Assemblies\netcoreapp2.0\Olive.Entities.dll</HintPath>
-    </Reference>
-  </ItemGroup>
-
-</Project>
