--- Services/Olive.Services.Integration/Package.nuspec
@@ -1,18 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<package xmlns="http://schemas.microsoft.com/packaging/2010/07/nuspec.xsd">
-  <metadata>
-    <id>Olive.Integration</id>
-    <version>1.0.7</version>
-    <title>Olive Integration (Service)</title>
-    <authors>Geeks Ltd</authors>
-    <projectUrl>https://github.com/Geeksltd/Olive</projectUrl>
-    <iconUrl>http://licensing.msharp.co.uk/Images/OliveComponent.png</iconUrl>
-    <copyright>Copyright ©2017 Geeks Ltd - All rights reserved.</copyright>
-    <description>Olive Framework</description>
-  </metadata>
-  <files>
-    <file src="..\..\@Assemblies\netcoreapp2.0\Olive.Services.Integration.dll" target="lib\netcoreapp2.0\" />
-    <file src="..\..\@Assemblies\netcoreapp2.0\Olive.Services.Integration.pdb" target="lib\netcoreapp2.0\" />
-    <file src="..\..\@Assemblies\netcoreapp2.0\Olive.Services.Integration.xml" target="lib\netcoreapp2.0\" />
-  </files>
-</package>
