+++ Services/Olive.Services.Testing/Package.nuspec
@@ -2,7 +2,7 @@
 <package xmlns="http://schemas.microsoft.com/packaging/2010/07/nuspec.xsd">
   <metadata>
     <id>Olive.Testing</id>
-    <version>1.0.15</version>
+    <version>1.0.16</version>
     <title>Olive Testing (Service)</title>
     <authors>Geeks Ltd</authors>
     <projectUrl>https://github.com/Geeksltd/Olive</projectUrl>
